refactored movieHelpers
